requirements.txt
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.11.0
beautifulsoup4==4.14.2
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
distro==1.9.0
docutils==0.22.2
ebooklib==0.19
faiss-cpu==1.12.0
filelock==3.20.0
fsspec==2025.9.0
google-auth==2.41.1
google-genai==1.39.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
id==1.5.0
idna==3.11
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
keyring==25.6.0
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
nh3==0.3.1
numpy==1.26.4
openai==1.82.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.12.3
pydantic-core==2.41.4
pygments==2.19.2
pypdf2==3.0.1
pyproject-hooks==1.2.0
python-dotenv==1.1.0
pyyaml==6.0.3
qrcode==8.2
readme-renderer==44.0
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==4.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
twine==6.2.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
websockets==15.0.1
azure-storage-blob==12.19.0
